+++ relm4-components/src/lib.rs
@@ -12,3 +12,4 @@ pub mod alert;
 pub mod open_button;
 pub mod open_dialog;
 pub mod save_dialog;
+pub mod simple_combo_box;