+++ CHANGES.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+
++ components: Add `SimpleComboBox` type as a more idiomatic wrapper around `gtk::ComboBoxText`
 + core: Append children for `gtk::Dialog` to its content area instead of using `set_child`
 + core: Remove `input` and `output` fields on `ComponentSender` and `FactoryComponentSender` in favor of `input_sender` and `output_sender` methods
 + core: Make `ComponentSender` and `FactoryComponentSender` structs instead of type aliases